requirements.txt
absl-py==1.4.0
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
array_record==0.5.1
astunparse==1.6.3
black==24.10.0
bleach==6.2.0
blinker==1.9.0
blis==1.1.0
builtin-interfaces==1.2.1
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
cmake==3.31.2
colorama==0.4.6
composition-interfaces==1.2.1
confection==0.1.5
contourpy==1.3.1
cssselect2==0.7.0
cv-bridge==3.2.1
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
defusedxml==0.7.1
demo-nodes-py==0.20.5
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@c69939aa85460e8135f40bce908a6cddaa73065f
diagnostic-msgs==4.2.4
dlib==19.24.6
dm-tree==0.1.8
docstring_parser==0.16
domain-coordinator==0.10.0
etils==1.11.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
Flask==3.1.0
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
fvcore==0.1.5.post20221221
gast==0.6.0
geometry-msgs==4.2.4
gin-config==0.5.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.68.1
gTTS==2.5.3
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.26.2
hydra-core==1.3.2
hyperframe==5.2.0
idna==2.10
image-geometry==3.2.1
immutabledict==4.2.1
importlib_resources==6.4.5
imutils==0.5.4
interactive-markers==2.3.2
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kaggle==1.6.17
keras==3.7.0
keras-nightly==3.7.0.dev2024122303
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
libclang==18.1.1
lifecycle-msgs==1.2.1
lit==18.1.8
logging-demo==0.20.5
lxml==5.3.0
map-msgs==2.1.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
message-filters==4.3.5
ml-dtypes==0.4.1
mpmath==1.3.0
murmurhash==1.0.11
mypy-extensions==1.0.0
namex==0.0.8
nav-msgs==4.2.4
networkx==3.4.2
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.13.1
osrf-pycommon==2.0.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.0.0
preshed==3.0.9
promise==2.3
proto-plus==1.25.0
protobuf==3.20.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.16.1
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-qt-binding==1.1.2
python-slugify==8.0.4
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.0
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
resource-retriever==3.1.2
rfc3986==1.5.0
rich==13.9.4
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
sentencepiece==0.2.0
seqeval==1.2.2
shape-msgs==4.2.4
shellingham==1.5.4
shiboken2==5.15.2.1
simple-parsing==0.1.6
six==1.17.0
slack_sdk==3.33.5
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
sros2==0.10.5
srsly==2.5.0
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.19.0a20241223
teleop-twist-keyboard==2.4.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-model-optimization==0.8.0
tensorflow-text-nightly==2.19.0.dev20241219
termcolor==2.5.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-hub-nightly==0.17.0.dev202412230315
tf-models-nightly==2.19.0.dev20241223
tf-slim==1.1.0
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
tf_keras-nightly==2.19.0.dev2024122310
tf_nightly==2.19.0.dev20241219
tflite-runtime==2.14.0
thinc==8.3.3
threadpoolctl==3.5.0
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.2.1
topic-monitor==0.20.5
torch==1.13.0
torchaudio==0.13.0+rocm5.2
torchvision==0.14.0+cu117
tqdm==4.66.6
trajectory-msgs==4.2.4
transformers==4.46.0
triton==2.0.0
turtlesim==1.4.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
unique-identifier-msgs==2.2.1
uritemplate==4.1.1
urllib3==2.3.0
visualization-msgs==4.2.4
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
yacs==0.1.8
zipp==3.21.0
